# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mission Control API",
    "description": "AI Agent Orchestration Dashboard — REST API for managing agents, tasks, sessions, tokens, webhooks, and more.",
    "version": "1.2.0",
    "license": { "name": "MIT" },
    "contact": { "name": "Mission Control", "url": "https://github.com/builderz-labs/mission-control" }
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "security": [
    { "SessionCookie": [] },
    { "ApiKey": [] }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc-session",
        "description": "Session token obtained via /api/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key (env var or DB-backed managed token)"
      }
    },
...
  },
  "paths": {
...
    "/api/webhooks/inbound": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Receive inbound webhook with signature verification",
        "description": "External services POST events here. Requires X-Webhook-Signature header with HMAC-SHA256.",
        "parameters": [{ "name": "slug", "in": "query", "required": true, "schema": { "type": "string" } }],
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "event": { "type": "string" }, "data": { "type": "object" } } } } } },
        "responses": {
          "200": { "description": "Event received" },
          "401": { "description": "Invalid or missing signature" },
          "404": { "description": "Webhook slug not found" }
        }
      }
    },
    "/api/webhooks/inbound/manage": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List inbound webhook configurations",
        "responses": { "200": { "description": "Inbound webhook list" } }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create inbound webhook",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "slug"], "properties": { "name": { "type": "string" }, "slug": { "type": "string" }, "allowed_events": { "type": "array", "items": { "type": "string" } } } } } } },
        "responses": { "201": { "description": "Inbound webhook created, secret shown once" } }
      },
      "put": {
        "tags": ["Webhooks"],
        "summary": "Update inbound webhook",
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Webhooks"],
        "summary": "Delete inbound webhook",
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
  },
...
}